parse fix

+++ main.js
@@ -1,6 +1,6 @@
 $(document).on("click", "button", function() {
     
-    if(parseInt($(this).attr("data-id")) || $(this).attr("data-id") == 0){
+    if(parseFloat($(this).attr("data-id")) || $(this).attr("data-id") == 0){
         console.log($(this).attr("data-id"));
         writeToInputField($(this).attr("data-id"));
     }
